demos/kafka-to-snowflake/data-generator/requirements.txt
flask>=3.0.0
kafka-python>=2.0.2
gunicorn>=21.0.0
